+++ test/hexapdf/layout/test_box_fitter.rb
@@ -1,31 +1,31 @@
 # -*- encoding: utf-8 -*-
 
 require 'test_helper'
-require 'hexapdf/layout/multi_frame'
+require 'hexapdf/layout/box_fitter'
 require 'hexapdf/layout'
 
-describe HexaPDF::Layout::MultiFrame do
+describe HexaPDF::Layout::BoxFitter do
   before do
     frames = [
       HexaPDF::Layout::Frame.new(0, 0, 100, 100),
       HexaPDF::Layout::Frame.new(100, 100, 50, 50),
     ]
-    @multi_frame = HexaPDF::Layout::MultiFrame.new(frames)
+    @box_fitter = HexaPDF::Layout::BoxFitter.new(frames)
   end
 
   def fit_box(count, width: 10, height: 10)
     ibox = HexaPDF::Layout::InlineBox.create(width: width, height: height) {}
-    @multi_frame.fit(HexaPDF::Layout::TextBox.new(items: [ibox] * count))
+    @box_fitter.fit(HexaPDF::Layout::TextBox.new(items: [ibox] * count))
   end
 
   def check_result(*pos, content_heights:, successful: true, boxes_remain: false)
     pos.each_slice(2).with_index do |(x, y), index|
-      assert_equal(x, @multi_frame.fit_results[index].x)
-      assert_equal(y, @multi_frame.fit_results[index].y)
+      assert_equal(x, @box_fitter.fit_results[index].x)
+      assert_equal(y, @box_fitter.fit_results[index].y)
     end
-    assert_equal(content_heights, @multi_frame.content_heights)
-    successful ? assert(@multi_frame.fit_successful?) : refute(@multi_frame.fit_successful?)
-    rboxes = @multi_frame.remaining_boxes.empty?
+    assert_equal(content_heights, @box_fitter.content_heights)
+    successful ? assert(@box_fitter.fit_successful?) : refute(@box_fitter.fit_successful?)
+    rboxes = @box_fitter.remaining_boxes.empty?
     boxes_remain ? refute(rboxes) : assert(rboxes)
   end
 
@@ -55,6 +55,6 @@ describe HexaPDF::Layout::MultiFrame do
     fit_box(20)
     check_result(0, 20, 0, 0, 100, 110, 100, 100, successful: false, boxes_remain: true,
                  content_heights: [100, 50])
-    assert_equal(2, @multi_frame.remaining_boxes.size)
+    assert_equal(2, @box_fitter.remaining_boxes.size)
   end
 end