+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Added
 
 * [HexaPDF::Layout::ColumnBox] for rendering content inside columns
-* [HexaPDF::Layout::MultiFrame] for placing boxes into  multiple frames
+* [HexaPDF::Layout::BoxFitter] for placing boxes into multiple frames
 * New configuration option 'debug' for enabling debug output
 * [HexaPDF::Document::Pages#move] for moving pages around the same document
 * [HexaPDF::Composer#box] for drawing arbitrary, registered boxes